+++ CHANGELOG.md
@@ -22,6 +22,8 @@ Releases marked with 🧪 (or previously with the "beta" suffix) were released o
 
 ### 2025.2.13 - 2025-10-30
 
+* 🔧 On Android 16 on Samsung devices, increase the limit of widget items back to 100.
+
 ### 2025.2.12 - 2025-09-19
 
 * 🔨 On Android 16 on Samsung devices, correctly update the widget on changes.